use interval moved out
