+++ src/additional-scss/_settings.scss
@@ -1,0 +1,2 @@
+$theme-font-path: "~uswds/dist/fonts";
+$theme-image-path: "~uswds/dist/img";
