--- src/components/AccordionItemTitle/AccordionItemTitle.css
@@ -1,17 +1,0 @@
-
-.rectBTN {
-
-    background-color: white;
-    border: solid;
-    color: grey;
-    padding: 10px 20px;
-    text-align: center;
-    text-decoration: none;
-    cursor: pointer;
-    outline:none;
-    height: 10px;
-    display :flex;
-    align-items: center;
-    float:right;
-    font-weight:bold;
-  }
